stream: extract Readable.from in its own file
See: https://github.com/nodejs/readable-stream/pull/420
